Data.DoubleMonoid.LZLC.Class

Documentation

class DMLZ a => DMLZLC a where

DoubleMonoid with left zero (DMLZ) and left catch

-- left catch
one /+/ x === one

Minimal complete definition

Nothing

Methods

msumZ :: ZList a -> a

(zero, /+/) is a monoid with a right zero element one. This can be stated as msumZ is a ZList algebra.

msumZ . fmap msumZ === msumZ . join

When you override the default implementation, it must keep the relation to the other methods:

msumZ Nend === zero
msumZ Zend === one
mprodZ (Cons a Nend) === a